requirements.txt
Django==2.1.6
django-bootstrap4==0.0.7
django-colorful==1.3
Pillow==5.4.1
pytz==2018.9
